+++ NEWS.md
@@ -9,12 +9,13 @@ The previous behavior "don't trigger an error if a bookmark is missing"
 is back. 
 - force Word default continuous section if no type is specified and first 
 page footer/header if it is seen.
-- enable images duplicated names when directories are not the same. 
 
 ## Changes
 
 - Change license to MIT.
 - drop internal `get_at_cursor()`
+- use images sha1 value for filename with package 'openssl'. This should fix 
+issues with duplicated images basenames, or non utf names, etc.
 
 ## Features